EAS and MTS - Clinical Director

+++ fe-main/src/main/java/com/ecquaria/cloud/moh/iais/service/impl/SyncAuditTrailRecordsServiceImpl.java
@@ -257,9 +257,11 @@ public class SyncAuditTrailRecordsServiceImpl implements SyncAuditTrailRecordsSe
             }
             return false;
         });
-        for(File f:files){
-            if(f.exists()&&f.isFile()){
-                MiscUtil.deleteFile(f);
+        if (files != null && files.length > 0) {
+            for (File f : files) {
+                if (f.exists() && f.isFile()) {
+                    MiscUtil.deleteFile(f);
+                }
             }
         }
         files = file.listFiles((dir, name) -> {
@@ -268,9 +270,11 @@ public class SyncAuditTrailRecordsServiceImpl implements SyncAuditTrailRecordsSe
             }
             return false;
         });
-        for(File f:files){
-            if(f.exists()&&f.isFile()){
-                MiscUtil.deleteFile(f);
+        if (files != null && files.length > 0) {
+            for (File f : files) {
+                if (f.exists() && f.isFile()) {
+                    MiscUtil.deleteFile(f);
+                }
             }
         }
     }
